Restrict char in the range of ASCII.
picrin accepts the code `(char->integer (integer->char 128))` but the
result is machine-dependent, since the C `char` type can be either
signed or unsigned as permitted by the spec. Unfortunately, the result
of signed one is incompatible with Unicode-aware implementations. For
interoperability, we should deny the range beyond ASCII for now.

Ref: #211

Signed-off-by: Masanori Ogino <masanori.ogino@gmail.com>